requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
pydantic>=2.7.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
python-dotenv